2021.02.18 아임포트적용 결제 API 구현

+++ src/main/java/com/example/demo/database/Entity/StudentEntity.java
@@ -25,7 +25,7 @@ public class StudentEntity implements Serializable {
     private String email;
     private String birth;
     private String address;
-    private int lecture;
+    private long lecture;
     private String gender;
     private String regdate;
